Split DetailsActivity and UploadActivity
