+++ app/src/main/res/layout-land/content_upload.xml
@@ -6,8 +6,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="de.bahnhoefe.deutschlands.bahnhofsfotos.DetailsActivity"
-    tools:showIn="@layout/activity_details">
+    tools:context="de.bahnhoefe.deutschlands.bahnhofsfotos.UploadActivity"
+    tools:showIn="@layout/activity_upload">
 
     <RelativeLayout
         android:layout_width="match_parent"
@@ -27,14 +27,6 @@
             android:indeterminate="true"
             android:visibility="gone" />
 
-        <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/viewPager"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_centerHorizontal="true"
-            android:layout_below="@+id/progressBar" />
-
-        <!--
         <com.github.chrisbanes.photoview.PhotoView
             android:id="@+id/imageview"
             android:layout_width="match_parent"
@@ -44,61 +36,36 @@
             android:src="@drawable/ic_stations_with_photo"
             android:scaleType="centerCrop"
             tools:scaleType="centerCrop"
-            android:contentDescription="@string/station_photo"/>
-            -->
+            android:contentDescription="@string/station_photo" />
 
-        <RelativeLayout
-            android:id="@+id/header"
-            android:layout_width="match_parent"
+        <EditText
+            android:id="@+id/etbahnhofname"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentTop="true"
             android:layout_alignParentStart="true"
-            android:layout_alignTop="@id/header">
-
-            <ImageView
-                android:id="@+id/marker"
-                android:layout_width="35sp"
-                android:layout_height="35sp"
-                android:src="@drawable/marker_missing"
-                android:scaleType="centerCrop"
-                android:layout_marginTop="8dp"
-                android:padding="8dp"
-                tools:scaleType="centerCrop"
-                android:contentDescription="@string/station_info"
-                android:onClick="showStationInfo" />
-
-            <EditText
-                android:id="@+id/etbahnhofname"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentTop="true"
-                android:layout_toEndOf="@id/marker"
-                android:textSize="25sp"
-                android:textStyle="bold"
-                android:textIsSelectable="true"
-                android:inputType="none"
-                android:background="#80ffffff"
-                android:padding="8dp"
-                android:hint="@string/station_name"
-                android:autofillHints="" />
-
-        </RelativeLayout>
+            android:textSize="25sp"
+            android:textStyle="bold"
+            android:textIsSelectable="true"
+            android:inputType="none"
+            android:background="#80ffffff"
+            android:padding="8dp"
+            android:hint="@string/station_name"
+            android:autofillHints="" />
 
         <TextView
-            android:id="@+id/license_tag"
+            android:id="@+id/upload_status"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentBottom="true"
-            android:layout_alignEnd="@id/viewPager"
-            android:text="@string/license_tag"
+            android:layout_alignEnd="@id/imageview"
             android:background="#80ffffff"
             android:padding="8dp"
             android:textColorLink="@color/colorPrimary" />
-
-
+        
         <LinearLayout
             android:id="@+id/buttons"
-            android:layout_above="@id/license_tag"
+            android:layout_above="@id/upload_status"
             android:layout_centerHorizontal="true"
             android:layout_centerVertical="true"
             android:layout_width="wrap_content"
@@ -119,14 +86,6 @@
                 android:src="@drawable/ic_photo_red_48px"
                 android:background="@null"
                 android:contentDescription="@string/select_photo" />
-
-            <ImageButton
-                android:id="@+id/button_report_problem"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/ic_bullhorn_48px"
-                android:background="@null"
-                android:contentDescription="@string/report_problem" />
 
             <ImageButton
                 android:id="@+id/button_upload"
