+++ app/src/main/res/menu/upload.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/share_photo"
+        android:orderInCategory="400"
+        android:icon="@drawable/ic_share_white_24px"
+        app:itemIconTint="@color/colorActions"
+        app:itemTextColor="@color/colorActions"
+        android:title="@string/share_photo"
+        app:showAsAction="ifRoom" />
+
+</menu>
